Move API metadata int Google.Cloud.Tools.Common
This will allow more tools to be written using the metadata.
