+++ tools/Google.Cloud.Tools.Common/Google.Cloud.Tools.Common.csproj
@@ -5,9 +5,13 @@
     <TargetFrameworks>netstandard1.5;net451</TargetFrameworks>
     <AssemblyName>Google.Cloud.Tools.Common</AssemblyName>
     <PackageId>Google.Cloud.Tools.Common</PackageId>
-    <NetStandardImplicitPackageVersion>1.6.0</NetStandardImplicitPackageVersion>
+    <NetStandardImplicitPackageVersion>1.6.1</NetStandardImplicitPackageVersion>
     <TreatWarningsAsErrors>True</TreatWarningsAsErrors>
   </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Newtonsoft.Json" Version="10.0.3" />
+  </ItemGroup>
 
   <ItemGroup Condition=" '$(TargetFramework)' == 'net451' ">
     <Reference Include="System" />
