+++ tools/Google.Cloud.Tools.GenerateDocfxSources/Google.Cloud.Tools.GenerateDocfxSources.csproj
@@ -17,7 +17,6 @@
 
   <ItemGroup>
     <PackageReference Include="YamlDotNet.Signed" Version="3.9.0" />
-    <PackageReference Include="Newtonsoft.Json" Version="9.0.1" />
   </ItemGroup>
 
 </Project>
