+++ tools/Google.Cloud.Tools.ProjectGenerator/Google.Cloud.Tools.ProjectGenerator.csproj
@@ -15,8 +15,4 @@
     <ProjectReference Include="..\Google.Cloud.Tools.Common\Google.Cloud.Tools.Common.csproj" />
   </ItemGroup>
 
-  <ItemGroup>
-    <PackageReference Include="Newtonsoft.Json" Version="10.0.2" />
-  </ItemGroup>
-
 </Project>
