+++ tools/Google.Cloud.Tools.ProjectGenerator/Program.cs
@@ -13,26 +13,28 @@
 // limitations under the License.
 
 using Google.Cloud.Tools.Common;
-using Newtonsoft.Json;
 using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using System.Xml.Linq;
 
 namespace Google.Cloud.Tools.ProjectGenerator
 {
     public class Program
     {
+        private const string GrpcVersion = "1.4.0";
+        private const string GaxVersion = "2.0.0";
+
+        private const string AnalyzersPath = @"..\..\..\tools\Google.Cloud.Tools.Analyzers\bin\$(Configuration)\netstandard1.3\publish\Google.Cloud.Tools.Analyzers.dll";
+        private const string StripDesktopOnNonWindows = @"..\..\..\StripDesktopOnNonWindows.xml";
+
         static int Main()
         {
             try
             {
                 var root = DirectoryLayout.DetermineRootDirectory();
-
-                var json = File.ReadAllText(Path.Combine(root, "apis", "apis.json"));
-                var apis = JsonConvert.DeserializeObject<List<ApiMetadata>>(json);
-
-                foreach (var api in apis)
+                foreach (var api in ApiMetadata.LoadApis())
                 {
                     GenerateProjects(Path.Combine(root, "apis", api.Id), api);
                 }
@@ -66,12 +68,12 @@ namespace Google.Cloud.Tools.ProjectGenerator
                 switch (suffix)
                 {
                     case "":
-                        api.GenerateMainProject(dir);
+                        GenerateMainProject(api, dir);
                         break;
                     case ".IntegrationTests":
                     case ".Tests":
                     case ".Snippets":
-                        api.GenerateTestProject(dir);
+                        GenerateTestProject(api, dir);
                         break;
                 }
             }
@@ -79,5 +81,173 @@ namespace Google.Cloud.Tools.ProjectGenerator
             // TODO: Solution file
             // TODO: Updates for unknown project types? Tricky...
         }
+
+        private static void GenerateMainProject(ApiMetadata api, string directory)
+        {
+            if (api.Version == null)
+            {
+                throw new Exception($"No version specified for {api.Id}");
+            }
+            string targetFrameworks = api.TargetFrameworks;
+            var dependencies = new SortedList<string, string>(api.Dependencies)
+            {
+                { "ConfigureAwaitChecker.Analyzer", "1.0.0-beta4" }
+            };
+            switch (api.Type)
+            {
+                case "rest":
+                    dependencies.Add("Google.Api.Gax.Rest", GaxVersion);
+                    targetFrameworks = targetFrameworks ?? "netstandard1.3;net45";
+                    break;
+                case "grpc":
+                    dependencies.Add("Google.Api.Gax.Grpc", GaxVersion);
+                    dependencies.Add("Grpc.Core", GrpcVersion);
+                    targetFrameworks = targetFrameworks ?? "netstandard1.5;net45";
+                    break;
+            }
+
+            var propertyGroup = new XElement("PropertyGroup",
+                // Build-related properties
+                new XElement("Version", api.Version), // TODO: Version, or VersionPrefix/VersionSuffix?
+                new XElement("TargetFrameworks", targetFrameworks),
+                new XElement("Features", "IOperation"),
+                new XElement("GenerateDocumentationFile", true),
+                new XElement("AssemblyOriginatorKeyFile", "../../GoogleApis.snk"),
+                new XElement("SignAssembly", true),
+                new XElement("Deterministic", true),
+                new XElement("PublicSign", new XAttribute("Condition", " '$(OS)' != 'Windows_NT' "), true),
+                new XElement("TreatWarningsAsErrors", true),
+                // Package-related properties
+                new XElement("Description", api.Description),
+                new XElement("PackageTags", string.Join(";", api.Tags.Concat(new[] { "Google", "Cloud" }))),
+                new XElement("IncludeSymbols", true),
+                new XElement("IncludeSource", true),
+                new XElement("Copyright", "Copyright 2017 Google Inc."),
+                new XElement("Authors", "Google Inc."),
+                new XElement("IconUrl", "https://cloud.google.com/images/gcp-icon-64x64.png"), // TODO: Check element name
+                new XElement("PackageLicenseUrl", "http://www.apache.org/licenses/LICENSE-2.0"),
+                new XElement("PackageProjectUrl", "https://github.com/GoogleCloudPlatform/google-cloud-dotnet"),
+                new XElement("RepositoryType", "git"),
+                new XElement("RepositoryUrl", "https://github.com/GoogleCloudPlatform/google-cloud-dotnet")
+            );
+            WriteProjectFile(api, directory, propertyGroup,
+                CreateDependenciesElement(dependencies, includeAnalyzers: true));
+        }
+
+        private static void GenerateTestProject(ApiMetadata api, string directory)
+        {
+            var dependencies = new SortedList<string, string>(api.TestDependencies)
+            {
+                { $@"..\{api.Id}\{api.Id}.csproj", "" }, // Main project
+                { "Microsoft.NET.Test.Sdk", "15.0.0" },
+                { "xunit", "2.3.0-beta1-build3642" },
+                { "xunit.runner.visualstudio", "2.3.0-beta1-build1309 " },
+                { "Moq", "4.7.8" }
+            };
+            var propertyGroup =
+                new XElement("PropertyGroup",
+                    new XElement("TargetFrameworks", api.TestTargetFrameworks ?? api.TargetFrameworks ?? "netcoreapp1.0;net452"),
+                    new XElement("IsPackable", false),
+                    new XElement("AssemblyOriginatorKeyFile", "../../GoogleApis.snk"),
+                    new XElement("SignAssembly", true),
+                    new XElement("PublicSign", new XAttribute("Condition", " '$(OS)' != 'Windows_NT' "), true),
+                    new XElement("TreatWarningsAsErrors", true),
+                    // 1701, 1702 and 1705 are disabled by default.
+                    // 4014 is required as snippets for streaming samples call Task.Run and don't await the result.
+                    // See https://github.com/googleapis/toolkit/issues/1271 - when that's fixed, we can remove this.
+                    new XElement("NoWarn", "1701;1702;1705;4014")
+                );
+            var itemGroup = CreateDependenciesElement(dependencies, includeAnalyzers: false);
+            // Allow test projects to use dynamic...
+            itemGroup.Add(new XElement("Reference",
+                new XAttribute("Condition", "'$(TargetFramework)' == 'net452'"),
+                new XAttribute("Include", "Microsoft.CSharp")));
+            // Test service... it keeps on getting added by Visual Studio, so let's just include it everywhere.
+            itemGroup.Add(new XElement("Service", new XAttribute("Include", "{82a7f48d-3b50-4b1e-b82e-3ada8210c358}")));
+            WriteProjectFile(api, directory, propertyGroup, itemGroup);
+        }
+
+        private static void WriteProjectFile(
+            ApiMetadata api, string directory, XElement propertyGroup, XElement dependenciesItemGroup)
+        {
+            var file = Path.Combine(directory, $"{Path.GetFileName(directory)}.csproj");
+            XElement doc;
+            // If the file already exists, load it and replace the elements (leaving any further PropertyGroup and ItemGroup elements).
+            // Make sure there's an appropriate import for stripping desktop builds on non-Windows platforms.
+            if (File.Exists(file))
+            {
+                doc = XElement.Load(file);
+                doc.Elements("Import").Where(x => (string)x.Attribute("Project") == @"..\..\StripDesktopOnNonWindows.xml").Remove();
+                doc.Elements("PropertyGroup").First().ReplaceWith(propertyGroup);
+                doc.Elements("ItemGroup").First().ReplaceWith(dependenciesItemGroup);
+
+                if (!doc.Elements("Import").Any(x => (string)x.Attribute("Project") == StripDesktopOnNonWindows))
+                {
+                    doc.Add(new XElement("Import", new XAttribute("Project", StripDesktopOnNonWindows)));
+                }
+            }
+            // Otherwise, create a new one
+            else
+            {
+                doc = new XElement("Project",
+                    new XAttribute("Sdk", "Microsoft.NET.Sdk"),
+                    propertyGroup,
+                    dependenciesItemGroup,
+                    new XElement("Import", new XAttribute("Project", StripDesktopOnNonWindows))
+                );
+            }
+
+            // Don't use File.CreateText as that omits the byte order mark.
+            // While byte order marks are nasty, Visual Studio will add it back any time a project file is
+            // manually edited, so it's best if we follow suit.
+            using (var stream = File.Create(Path.Combine(directory, $"{Path.GetFileName(directory)}.csproj")))
+            {
+                doc.Save(stream);
+            }
+        }
+
+        private static string GenerateProjectReference(string key)
+        {
+            // There are a few special cases which don't fit in, then the common case
+            // where it follows the normal pattern, then being flexible for anything
+            // that doesn't fit...
+            switch (key)
+            {
+                case "Google.Cloud.ClientTesting": return @"..\..\..\tools\Google.Cloud.ClientTesting\Google.Cloud.ClientTesting.csproj";
+                case var _ when !key.Contains(".csproj"): return $@"..\..\{key}\{key}\{key}.csproj";
+                default: return key;
+            }
+        }
+
+        // Dependencies with an empty value will be treated as project dependencies;
+        // dependencies with a value will be treated as package dependencies with the value as the version.
+        private static XElement CreateDependenciesElement(IDictionary<string, string> dependencies, bool includeAnalyzers) =>
+            new XElement("ItemGroup",
+                // Use the GAX version for all otherwise-unversioned GAX dependencies
+                dependencies
+                    .Where(d => d.Value == "" && d.Key.StartsWith("Google.Api.Gax"))
+                    .Select(d => new XElement("PackageReference",
+                        new XAttribute("Include", d.Key),
+                        new XAttribute("Version", GaxVersion))),
+                dependencies
+                    .Where(d => d.Value == "" && !d.Key.StartsWith("Google.Api.Gax"))
+                    .Select(d => new XElement("ProjectReference", new XAttribute("Include", GenerateProjectReference(d.Key)))),
+                dependencies
+                    .Where(d => d.Value != "")
+                    .Select(d => new XElement("PackageReference",
+                        new XAttribute("Include", d.Key),
+                        new XAttribute("Version", d.Value),
+                        // Make references to Grpc.Core deploy native dependencies
+                        // See https://github.com/GoogleCloudPlatform/google-cloud-dotnet/issues/1066
+                        d.Key == "Grpc.Core" ? new XAttribute("PrivateAssets", "None") : null,
+                        // Make references to ConfigureAwaitChecker effectively private
+                        d.Key == "ConfigureAwaitChecker.Analyzer" ? new XAttribute("PrivateAssets", "All") : null)
+                    ),
+                includeAnalyzers ?
+                    new XElement("Analyzer",
+                        new XAttribute("Condition", $"Exists('{AnalyzersPath}')"),
+                        new XAttribute("Include", AnalyzersPath)) :
+                    null
+            );
     }
 }
